in the middle of adding ray intersection code from my previous project

+++ MM3DECS/Shaders/Common.h
@@ -52,8 +52,8 @@ typedef struct{
 } LocalRay;
 
 typedef struct {
-  uint width;
-  uint height;
+  float width;
+  float height;
   uint tiling;
   uint lightCount;
   vector_float3 cameraPosition;
